sample*.json: Remove W+jets MC.
Since Wjets is treated as part of the NRB estimate and there are no
Wjets events in the emu category, the removal of the MC will not change
the NRB estimate. However, since there is a single high-weight event in
the ee channel, the uncertainty as calculated from the closure test
would be unrealistically inflated by the MC. Thus, it is removed.