runWIMP/Plotter: Split QCDScale/PDF Into norm. and accept.
The plotter rescales the up/down shapes variations of PDF and QCD to
match the cross-section of the nominal weighting.
The information for this comes from the cutflow histogram in runWIMP.
The plotter decides if a sample is a signal by parsing its name, i.e.
checking if there's "DM_V","DM_A"... etc in the name.